refactor:migrate topicPageCypressTests to NextJS
