--- cypress/support/helpers/getPathWithSuffix.ts
@@ -1,5 +1,0 @@
-export default ({ path, suffix = '' }) => {
-  const { pathname, search } = new URL(`https://www.bbc.com${path}`);
-
-  return `${pathname}${suffix}${search}`;
-};
