+++ ws-nextjs-app/cypress/e2e/topicPage/index.cy.ts
@@ -1,22 +1,24 @@
-import runTestsForPage from '#nextjs/cypress/support/helpers/runTestsForPage';
+import { TOPIC_PAGE } from '#app/routes/utils/pageTypes';
 import crossPlatformTests from './tests';
-import { TOPIC_PAGE } from '../../../../src/app/routes/utils/pageTypes';
-import urlValidationTest from '../../../support/helpers/urlValidationTest';
+import urlValidationTest from '../../support/helpers/urlValidationTest';
 import testsForAllCanonicalPages from '../testsForAllCanonicalPages';
-import getPathWithSuffix from '../../../support/helpers/getPathWithSuffix';
-import { assertLiteSiteSummaryComponentToMainSiteClick } from '../../specialFeatures/atiAnalytics/assertions/liteSiteSummary';
-import { setUserIDCookie } from '../../specialFeatures/atiAnalytics/helpers';
-import { assertPageView } from '../../specialFeatures/atiAnalytics/assertions';
+import getPathWithSuffix from '../../support/helpers/getPathWithSuffix';
+import { assertLiteSiteSummaryComponentToMainSiteClick } from '../specialFeatures/atiAnalytics/assertions/liteSiteSummary';
+import { setUserIDCookie } from '../specialFeatures/atiAnalytics/helpers';
+import { assertPageView } from '../specialFeatures/atiAnalytics/assertions';
 import {
   assertDropdownNavigationComponentClick,
   assertDropdownNavigationComponentView,
   assertScrollableNavigationComponentClick,
   assertScrollableNavigationComponentView,
-} from '../../specialFeatures/atiAnalytics/assertions/navigation';
+} from '../specialFeatures/atiAnalytics/assertions/navigation';
 import {
   assertMessageBannerComponentClick,
   assertMessageBannerComponentView,
-} from '../../specialFeatures/atiAnalytics/assertions/messageBanner';
+} from '../specialFeatures/atiAnalytics/assertions/messageBanner';
+import runTestsForPage, {
+  TestDataType,
+} from '../../support/helpers/runTestsForPage';
 
 const tests = [
   crossPlatformTests,
@@ -111,7 +113,7 @@ const atiAnalyticsTestSuites = [
       assertMessageBannerComponentClick,
     ],
   },
-];
+] as unknown as TestDataType[];
 
 const atiAnalyticsLiteTestSuites = atiAnalyticsTestSuites.map(testSuite => {
   const excludedLiteTests = [
